requirement.txt
langchain
langchain-ollama
langchain-huggingface
langchain-chroma
sentence-transformers
tiktoken
langfuse
datasets
python-dotenv